agent-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.11.0
python-dotenv==1.0.1
google-generativeai==0.8.0
httpx==0.28.0
pydantic==2.10.0
